stok kartı cins özellikleri tablosu ayrıldı.
cins özellikleri tablosunda silme işlemini db dahi yapamıyorum. Neden?????? kontrol edilecek.

+++ ERP/BackEnd/Core/Ths.Database.Table.pas
@@ -787,7 +787,7 @@ begin
     Close;
     SQL.Clear;
     SQL.Text := 'DELETE FROM ' + TableName + ' WHERE id=:id;';
-    ParamByName(Id.FieldName).Value := Id.AsInteger;
+    ParamByName(Id.FieldName).Value := Id.AsInt64;
     ExecSQL;
     Close;
   finally
